pyproject.toml
[tool.poetry]
name = "sql-training"
version = "0.1.0"
description = ""
authors = ["Carlos Huaccho <Carlos@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
ipython = "^8.27.0"
sqlalchemy = "^2.0.34"
pylint = "^3.2.7"
flake8 = "^7.1.1"
black = "^24.8.0"
pytest = "^8.3.2"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.9"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
